+++ packages/pluggableWidgets/combobox-web/src/__tests__/StaticSelection.spec.tsx
@@ -43,6 +43,7 @@ describe("Combo box (Static values)", () => {
             clearable: true,
             filterType: "contains",
             selectedItemsStyle: "text",
+            readOnlyStyle: "bordered",
             clearButtonAriaLabel: dynamicValue("Clear selection"),
             removeValueAriaLabel: dynamicValue("Remove value"),
             selectAllButtonCaption: dynamicValue("Select All"),