+++ packages/pluggableWidgets/combobox-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   We added readonly style to the combobox configuration.
+
 ## [1.4.0] - 2024-04-19
 
 ### Changed